[State Sync] Move StateSyncClient into own file.
